split out loading of example assets
