+++ app.js
@@ -1,4 +1,4 @@
-import { caveLion, lightbearer, berserker, ironclad } from './src/exampleAssets.js'
+import { loadDefaultExampleAsset } from './src/exampleAssetLoader.js'
 import { createAssetHtml, setSvgDimensions } from './src/asset.js'
 import { showScreen } from './src/router.js'
 import { assetScale } from './src/assetScaling.js'
@@ -7,15 +7,6 @@ import './src/download.js'
 const assetInput = document.querySelector('.interface-input')
 const assetContainer = document.querySelector(".assets")
 const updateButton = document.querySelector(".update")
-const showLightbearerAssetButton = document.querySelector("#lightbearer-example")
-const showIroncladAssetButton = document.querySelector("#ironclad-example")
-const showCaveLionAssetButton = document.querySelector("#cave-lion-example")
-
-const showSingleAssetExample = (asset) => {
-    assetInput.value = JSON.stringify(asset, null, 2)
-    assetContainer.innerHTML = createAssetHtml(asset, assetScale)
-    setSvgDimensions()
-}
 
 const showHelpButton = document.querySelector('#show-help')
 showHelpButton.onclick = () => showScreen("help")
@@ -32,9 +23,4 @@ updateButton.onclick = () => {
     setSvgDimensions()
 }
 
-
-showLightbearerAssetButton.onclick = () => showSingleAssetExample(lightbearer)
-showIroncladAssetButton.onclick = () => showSingleAssetExample(ironclad)
-showCaveLionAssetButton.onclick = () => showSingleAssetExample(caveLion)
-
-showSingleAssetExample(ironclad)
+loadDefaultExampleAsset()
